[Refactor]: WeeklyMentalUiState import 경로 model 패키지로 정리
