+++ app/src/main/java/com/konkuk/medicarecall/ui/statistics/weeklycard/WeeklyGlucoseCard.kt
@@ -22,8 +22,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.konkuk.medicarecall.ui.homedetail.glucoselevel.component.GlucoseStatusChip
-import com.konkuk.medicarecall.ui.statistics.WeeklyGlucoseUiState
+import com.konkuk.medicarecall.ui.statistics.model.WeeklyGlucoseUiState
 import com.konkuk.medicarecall.ui.theme.LocalMediCareCallShadowProvider
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
 import com.konkuk.medicarecall.ui.theme.figmaShadow
