+++ app/src/main/java/com/konkuk/medicarecall/ui/statistics/weeklycard/WeeklyMedicineCard.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.konkuk.medicarecall.R
-import com.konkuk.medicarecall.ui.statistics.WeeklyMedicineUiState
+import com.konkuk.medicarecall.ui.statistics.model.WeeklyMedicineUiState
 import com.konkuk.medicarecall.ui.statistics.WeeklySummaryUtil
 import com.konkuk.medicarecall.ui.theme.LocalMediCareCallShadowProvider
 import com.konkuk.medicarecall.ui.theme.MediCareCallTheme
